pyproject.toml

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "recordex"
version = "0.0.0"

dependencies = [
    # Django
    "Django==5.0.1",
    "django-polymorphic==3.1.0",
    "djangorestframework==3.14.0",
    "django_model_utils==4.3.1",
    "djangorestframework_simplejwt==5.3.1",
    # DB
    "psycopg-binary==3.1.17",
    "psycopg-pool==3.2.1",
    "psycopg==3.1.17",
    # Misc
    "python-dotenv==1.0.1",
]


[project.optional-dependencies]
test = [
    "pre-commit==3.6.0",
    "black==23.12.1",
    "pytest==7.4.4",
    "pytest-django==4.7.0",
    "pytest-mock==3.12.0",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
requires-python = ">=3.12"
python_files = ["test_*.py", "conftest.py"]

[tool.setuptools]
py-modules = []
